==> Fig2_RNA-seq/03_GSEA_edgeR.Rmd <==
---
title: "GSEA Analysis of edgeR analyzed genes"
author: "Mikhail Dozmorov"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: no
  html_document:
    theme: cerulean
    toc: yes
---

```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
# Set up the environment
library(knitr)
opts_chunk$set(cache.path='cache/', fig.path='img/', cache=F, tidy=T, fig.keep='high', echo=F, dpi=100, warnings=F, message=F, comment=NA, warning=F, results='as.is', fig.width = 10, fig.height = 6) #out.width=700, 
library(pander)
panderOptions('table.split.table', Inf)
set.seed(1)
library(dplyr)
options(stringsAsFactors = FALSE)
```

```{r libraries}
library(readxl)
library(writexl)
library(enrichR) # install.packages("enrichR")
library(annotables) # devtools::install_github("stephenturner/annotables") Annotations of Ensembl IDs
library(clusterProfiler)
library(DOSE)
library(ggplot2)
library(readr)
library(HGNChelper)
library(tidyr)
```

# Settings

```{r settings}
# Which pathway enrichment analysis to run
run_gsea <- TRUE # If TRUE, GSEA pathway enrichment analysis is run, otherwise, standard hypergeometric-based enrichment
min_kegg_genes <- 20 # Minimum number of genes to run enrichment analysis on
max_kegg_genes <- 4000 # Maximum number of genes to run enrichment analysis on
up_dn_separate <- FALSE # Whether to run KEGG separately on up- and downregulated genes. FALSE - do not distinguish directionality
FDR_cutoff     <- 0.1 # FDR cutoff for DEGs
logFC_cutoff   <- 0 # Absolute logFC cutoff, 0 for no filtering
p_adj_cutoff   <- 1 # FDR cutoff for enrichment results
nperm          <- 1000 # Number of permutations (for GSEA)
# Sort genes by
pvalue <- TRUE # If FALSE (default), genes are sorted by log2FC. If TRUE, sorted by -log10(p-value)

# Data
data_dir       <- "/Users/mdozmorov/Documents/Data/GoogleDrive/HiC_files/results/RNA-seq/2021.11.19_UCD52_DEG" # Working directory
fileNameIn1   <- file.path(data_dir, "data/Data_Human_DESeq2_UCD52_DEGs_Untreated.vs.CarboplatinResistant_04.06.22.txt") # File to save all results of differential expression analysis
# degs_sheet     <- "1.protein-coding" # Which worksheet contains differentially expressed genes
fileNameOut1 <- file.path(data_dir, paste0("results/04.06.22_GSEA_edgeR_UCD52PR_CR_", FDR_cutoff, "_", logFC_cutoff, ifelse(pvalue, "_pvalue", ""), "1.xlsx"))
fileNameOut2 <- file.path(data_dir, "results/04.06.22_DEGs_edgeR_UCD52PR_CR_annotated.xlsx")
```

```{r}
# Mutually exclusive selectors, which analysis should be run. Only one may be TRUE
human_analysis <- TRUE
mouse_analysis <- FALSE

# Prepate gene annotations, remove non-canonical chromosome names
if (human_analysis) {
  library(org.Hs.eg.db)
  OrgDb = "org.Hs.eg.db"; species = "hsa"
  gene_annotations <- grch38[ !(grepl("_", grch38$chr) | grepl("GL", grch38$chr)), c("ensgene", "symbol", "biotype", "description")]
  # gene_annotations <- gene_annotations[ !duplicated(gene_annotations) & !is.na(gene_annotations$symbol) & gene_annotations$description != "", ]
  KEGG <- "KEGG_2019_Human"
  msigdbr_org <- "Homo sapiens"
  # Gene length for TPM calculations
  gene_length <- data.frame(Geneid = grch38$ensgene, Length = grch38$end - grch38$start)
}
if (mouse_analysis) {
  library(org.Mm.eg.db)
  OrgDb = "org.Mm.eg.db"; species = "mmu"
  gene_annotations <- grcm38[ !(grepl("_", grcm38$chr) | grepl("GL", grcm38$chr)), c("ensgene", "symbol", "biotype", "description")]
  gene_annotations <- gene_annotations[ !duplicated(gene_annotations) & !is.na(gene_annotations$symbol) & gene_annotations$description != "", ]
  KEGG <- "KEGG_2019_Mouse"
  msigdbr_org <- "Mus musculus"
  # Gene length for TPM calculations
  gene_length <- data.frame(Geneid = grcm38$ensgene, Length = grcm38$end - grcm38$start)
}
# All genes for background
all.symbol <- unique(gene_annotations$symbol) 
```

# Load data

```{r}
mtx <- read_tsv(fileNameIn1)
# Reformat columns
# mtx <- mtx[!is.na(mtx$symbol), ]
mtx$ensgenes <- mtx$genes
mtx$genes <- mtx$symbol
mtx <- mtx[, c("ensgenes", "genes", "logFC", "logCPM", "LR", "PValue", "FDR")]

mtx_nonna <- mtx[!is.na(mtx$genes), ]
mtx_na <- mtx[is.na(mtx$genes), ]

# Select min p-value
mtx_nonna <- mtx_nonna %>% group_by(genes) %>% slice_min(PValue) 
# Correct gene names
current_map <- getCurrentHumanMap()
genes_checked <- checkGeneSymbols(mtx_nonna$genes, map = current_map)
mtx_nonna$genes <- genes_checked$Suggested.Symbol
# Remove empty symbols
# mtx <- mtx[!is.na(mtx$genes), ]
# Add NAs
mtx <- rbind(mtx_nonna, mtx_na)
# Sort by p-value
mtx <- mtx[order(mtx$PValue, decreasing = FALSE), ]
# Annotate
gene_annotations <- gene_annotations[!duplicated(gene_annotations[, c("ensgene", "symbol", "biotype", "description")]), c("ensgene", "symbol", "biotype", "description")]
# Remove empty gene symbol records
# gene_annotations <- gene_annotations[gene_annotations$symbol != "", ]
# If a gene has multiple biotypes, take the first one
# gene_annotations <- gene_annotations %>% group_by(symbol, biotype) %>% slice_head()

mtx_annotated <- left_join(mtx, gene_annotations, by = c("ensgenes" = "ensgene"))
# Assign non-annotated gene symbold as undefined
mtx_annotated$biotype[is.na(mtx_annotated$biotype)] <- "undefined"
# Summary counts
# Upregulated genes
mtx_summary_up <- mtx_annotated$biotype[mtx_annotated$FDR < FDR_cutoff & mtx_annotated$logFC > 0] %>% table() %>% as_tibble()
colnames(mtx_summary_up) <- c("biotype", "count")
# Downregulated genes
mtx_summary_dn <- mtx_annotated$biotype[mtx_annotated$FDR < FDR_cutoff & mtx_annotated$logFC < 0] %>% table() %>% as_tibble()
colnames(mtx_summary_dn) <- c("biotype", "count")
# Combine them
mtx_summary_updn <- full_join(mtx_summary_up, mtx_summary_dn, by = "biotype", suffix = c(".up", ".dn"))
# Replace NAs with 0s
mtx_summary_updn[is.na(mtx_summary_updn)] <- 0

# Test whether biotype distribution is different
biotype_pval <- vector(length = length(mtx_summary_updn$biotype))
biotypes <- mtx_summary_updn$biotype
# For each biotype
for (i in 1:length(mtx_summary_updn$biotype)) {
  # 2x2 contingency table, target biotype vs. rest
  mtx_2x2 <- rbind(mtx_summary_updn[ mtx_summary_updn$biotype == biotypes[i], c("count.up", "count.dn")],
                 mtx_summary_updn[ mtx_summary_updn$biotype != biotypes[i], c("count.up", "count.dn")] %>% colSums(na.rm = TRUE))
  # Prop.test results are equivalent
  biotype_pval[i] <- chisq.test(mtx_2x2)$p.value
}
# Format p-values
mtx_summary_updn$p.value <- formatC(biotype_pval, format = "e", digits = 3)
# Add total
mtx_summary_updn <- rbind(mtx_summary_updn,
                          c("Total", sum(mtx_summary_updn$count.up, na.rm = TRUE), sum(mtx_summary_updn$count.dn, na.rm = TRUE), ""))

# Save annotated data
if (!file.exists(fileNameOut2)) {
  # Separate into protein-coding and others
  x <- list(protein.coding = mtx_annotated[mtx_annotated$biotype == "protein_coding", ],
            other = mtx_annotated[mtx_annotated$biotype != "protein_coding", ],
            DEGs = mtx_annotated %>% dplyr::filter(FDR < FDR_cutoff) %>% dplyr::arrange(biotype, genes, desc(PValue)),
            summary = mtx_summary_updn)
  write_xlsx(x, fileNameOut2)
}

# Remove NAs 
mtx_annotated <- mtx_annotated[!is.na(mtx_annotated$FDR) & mtx_annotated$biotype == "protein_coding" & !is.na(mtx_annotated$biotype) & !is.na(mtx_annotated$genes), ]
res <- data.frame(symbol = mtx_annotated$genes, logFC = mtx_annotated$logFC, pval = mtx_annotated$PValue, p.adj = mtx_annotated$FDR)
res <- res[res$p.adj < FDR_cutoff & abs(res$logFC) > logFC_cutoff, ]
res <- res[order(res$logFC, decreasing = TRUE), ]
```

# Overview

- All up- and downregulated genes are analyzed. It is possible to run the analysis only on up- or downregulated genes.
- All enrichment results are filtered at the adjusted p-value cutoff `r p_adj_cutoff`. This is the lowest minimum when one can say something is significant.

- Two types of Enrichment analyses were run:
    - **Enrich** - enrichment _without considering_ DEG order (Most frequently used, as in IPA)
    - **GSEA** - enrichment _considering_ largest-to-smallest order of DEGs ()

- The following signatures were used:
    - **KEGG** - KEGG pathways
    - MSigDB collections, see  https://www.gsea-msigdb.org/gsea/msigdb/index.jsp. 
        - **H**	hallmark gene sets  are coherently expressed signatures derived by aggregating many MSigDB gene sets to represent well-defined biological states or processes.
        - **C1**	positional gene sets  for each human chromosome and cytogenetic band.
        - **C2**	curated gene sets  from online pathway databases, publications in PubMed, and knowledge of domain experts.
        - **C3**	regulatory target gene sets  based on gene target predictions for microRNA seed sequences and predicted transcription factor binding sites.
        - **C4**	computational gene sets  defined by mining large collections of cancer-oriented microarray data.
        - **C5**	GO gene sets  consist of genes annotated by the same GO terms.
        - **C6**	oncogenic gene sets  defined directly from microarray gene expression data from cancer gene perturbations.
        - **C7**	immunologic gene sets  defined directly from microarray gene expression data from immunologic studies.
- Sheets in the Excel results file are named by the **type of analysis**.**signature**, e.g., **Enrich.KEGG**, **GSEA.C7**
    - Signatures in question can be googled

**KEGG pathway Legend:** "ID", "Description" - KEGG pathway ID/description, respectively; "NES" - [normalized enrichment score](http://software.broadinstitute.org/gsea/doc/GSEAUserGuideFrame.html); "pvalue", "p.adjust" - raw and FDR-adjusted p-values, respectively; "core_enrichment" - genes enriched in the corresponding pathway.

```{r}
  websiteLive <- TRUE # Check if EnrichR is up
  dbs <- listEnrichrDbs()
  if (is.null(dbs)) websiteLive <- FALSE
  if(websiteLive) {
    # Subset the number of DEGs for KEGG analysis to the maximum
    if (nrow(res) > max_kegg_genes) {
      degs_subset <- res[1:max_kegg_genes, ]
    } else {
      degs_subset <- res
    }
    # Get list of up- and downregulated genes
    up.genes <- sort(unique(degs_subset$symbol[ degs_subset$logFC > 0 ]))
    dn.genes <- sort(unique(degs_subset$symbol[ degs_subset$logFC < 0 ]))
    # Run KEGG
    if (up_dn_separate) {
      # Analyze up- and downregulated genes separately
      print(paste0("KEGG pathway run on ", length(up.genes), " upregulated and ", length(dn.genes), " downregulated genes."))
      # res.kegg <- save_enrichr(up.genes = up.genes, dn.genes = dn.genes, databases = KEGG, fdr.cutoff = p_adj_cutoff, fileName = fileName_rna, wb = wb, sheetName = "KEGG_1")
      res.kegg    <- NULL # Initially, empty value
      res.kegg.up <- enrichr(up.genes, databases = KEGG)
      res.kegg.dn <- enrichr(dn.genes, databases = KEGG)
      # If significant results are present, save them
      if(nrow(res.kegg.up[[KEGG]]) > 0 & sum(res.kegg.up[[KEGG]]$Adjusted.P.value < p_adj_cutoff) > 0) {
        res.kegg.up <- as.data.frame(res.kegg.up[[KEGG]])
        res.kegg.up <- res.kegg.up[res.kegg.up$Adjusted.P.value < p_adj_cutoff, , drop = FALSE]
        res.kegg.up <- res.kegg.up %>% mutate(Direction = "UP")
        res.kegg    <- rbind(res.kegg, res.kegg.up)
      }
      if(nrow(res.kegg.dn[[KEGG]]) > 0 & sum(res.kegg.dn[[KEGG]]$Adjusted.P.value < p_adj_cutoff) > 0) {
        res.kegg.dn <- as.data.frame(res.kegg.dn[[KEGG]])
        res.kegg.dn <- res.kegg.dn[res.kegg.dn$Adjusted.P.value < p_adj_cutoff, , drop = FALSE]
        res.kegg.dn <- res.kegg.dn %>% mutate(Direction = "DN")
        res.kegg    <- rbind(res.kegg, res.kegg.dn)
      }
    } else {
      # Analyze up- and downregulated genes together
      print(paste0("KEGG pathway run on ", length(unique(c(up.genes, dn.genes))), " genes without distinguishing them by directionality."))
      # res.kegg <- MDmisc::save_enrichr(up.genes = unique(c(up.genes, dn.genes)), databases = KEGG, fdr.cutoff = p_adj_cutoff, fileName = fileName_rna, wb = wb, sheetName = "KEGG_1")
      res.kegg <- enrichr(unique(c(up.genes, dn.genes)), databases = KEGG) # KEGG results only
      # If significant results are present, save them
      if(nrow(res.kegg[[KEGG]]) > 0 & sum(res.kegg[[KEGG]]$Adjusted.P.value < p_adj_cutoff) > 0) {
        res.kegg <- as.data.frame(res.kegg[[KEGG]])
        res.kegg <- res.kegg[res.kegg$Adjusted.P.value < p_adj_cutoff, , drop = FALSE]
      }
    }
  }
```

# GSEA analysis 

```{r}
  # All DEGs for GSEA
  res.all <- res
  # Convert symbols to entrezids
  eid <- bitr(res.all$symbol, fromType="SYMBOL", toType="ENTREZID", OrgDb=OrgDb)
  # Attach converted entrezids
  res.all <- left_join(res.all, eid, by = c("symbol" = "SYMBOL"))
  res.all <- res.all[ !is.na(res.all$ENTREZID), ]
  # List of t-statistics
  if (pvalue) {
    geneList <- -log10(res.all$pval) * sign(res.all$logFC)
  } else {
    geneList <- res.all$logFC
  }
  # Make it named
  names(geneList) <- res.all$ENTREZID
  # And decreasing sorted
  geneList <- sort(geneList, decreasing = TRUE)
  
  # Top DEGs for enrichr
  res.all <- res[res$p.adj < FDR_cutoff, ]
  # Convert symbols to entrezids
  eid <- bitr(res.all$symbol, fromType="SYMBOL", toType="ENTREZID", OrgDb=OrgDb)
  # Attach converted entrezids
  res.all <- left_join(res.all, eid, by = c("symbol" = "SYMBOL"))
  res.all <- res.all[ !is.na(res.all$ENTREZID), ]
  # List of t-statistics
  if (pvalue) {
    geneList_significant <- -log10(res.all$pval) * sign(res.all$logFC)
  } else {
    geneList_significant <- res.all$logFC
  }
  # Make it named
  names(geneList_significant) <- res.all$ENTREZID
  # And decreasing sorted
  geneList_significant <- sort(geneList_significant, decreasing = TRUE)

```

```{r}
  ## GSEA using clusterProfiler
  # Actual GSEA
  set.seed(1)
  ego3 <- gseKEGG(geneList     = geneList,
                  organism     = species,
                  # nPerm        = nperm,
                  minGSSize    = 10,
                  pvalueCutoff = 1,
                  verbose      = FALSE)
  # Get summary
  ego3 <- setReadable(ego3, OrgDb = OrgDb, keyType = "ENTREZID")
  res.kegg.gsea <- as.data.frame(ego3)
  # Prepare for table output
  res.kegg.gsea <- res.kegg.gsea[, c("ID", "Description", "NES", "pvalue", "p.adjust", "core_enrichment")]
  res.kegg.gsea <- res.kegg.gsea[order(abs(res.kegg.gsea$NES), decreasing = TRUE), ]
  res.kegg.gsea <- res.kegg.gsea[res.kegg.gsea$p.adjust < p_adj_cutoff, ]
  res.kegg.gsea$NES       <- round(res.kegg.gsea$NES, digits = 2)
  res.kegg.gsea$pvalue    <- formatC(res.kegg.gsea$pvalue, format = "e", digits = 2)
  res.kegg.gsea$p.adjust  <- formatC(res.kegg.gsea$p.adjust, format = "e", digits = 2)
  rownames(res.kegg.gsea) <- NULL
```

# MSigDB analysis

```{r}
# Convert Entrez IDs back to symbols
entrez_to_symbols <- function(x = res.msigdf.em$geneID) {
  sapply(x, function(xx) {
    xx_split <- strsplit(xx, "/")[[1]]
    # Convert entrezids to symbols
    eid <- bitr(xx_split, fromType="ENTREZID", toType="SYMBOL", OrgDb=OrgDb)
    # combine back
    paste(sort(eid$SYMBOL), collapse = "/")
  })
}

library(msigdbr)
m_df <- msigdbr(species = "Homo sapiens")
# table(m_df$gs_cat)
# 
m_df_gs_cat <- unique(m_df$gs_cat) %>% sort()
# m_df_gs_cat <- c("C2", "C5", "H")
res.msigdf.all <- list()
# gs_cat="H"
for (gs_cat in m_df_gs_cat) {
    # Term to gene
     m_t2g <- msigdbr(species = msigdbr_org, category = gs_cat) %>% 
       dplyr::distinct(gs_name, entrez_gene)
    # Term to description 
    m_t2d <- msigdbr(species = msigdbr_org, category = gs_cat) %>% 
       dplyr::distinct(gs_name, gs_description)
    # Enrichment analysis
  em <- enricher(names(geneList_significant), TERM2GENE=m_t2g, pvalueCutoff = p_adj_cutoff)
  res.msigdf.em <-em@result
  res.msigdf.em$geneID <- entrez_to_symbols(x = res.msigdf.em$geneID)
        res.msigdf.em <- res.msigdf.em %>% 
      # separate the rows by splitting by the delimiter and expanding the rows 
      separate_rows(geneID, convert = TRUE, sep = "/") %>% 
      # group the genes by their other columns to keep them 
      group_by(ID, Description, GeneRatio, BgRatio, pvalue, p.adjust, qvalue, Count) %>%
      # sort the genes for each term 
      arrange(geneID) %>% 
      summarise(geneID = paste(geneID, collapse="/")) %>% 
      arrange(pvalue, p.adjust)
      # Append description
      res.msigdf.em <- left_join(res.msigdf.em, m_t2d, by = c("ID" = "gs_name"))
      res.msigdf.em$Description <- res.msigdf.em$gs_description

  em2 <- GSEA(geneList, TERM2GENE = m_t2g, pvalueCutoff = p_adj_cutoff)
  res.msigdf.em2 <-em2@result
  res.msigdf.em2$core_enrichment <- entrez_to_symbols(x = res.msigdf.em2$core_enrichment)
        # sort the genes alphabetically
      res.msigdf.em2 <- res.msigdf.em2 %>% 
      # separate the rows by splitting by the delimiter and expanding the rows 
      separate_rows(core_enrichment, convert = TRUE, sep = "/") %>% 
      # group the genes by their other columns to keep them 
      group_by(ID, Description, setSize, enrichmentScore, NES, pvalue, p.adjust, qvalues,
               rank, leading_edge) %>%
      # sort the genes for each term 
      arrange(core_enrichment) %>% 
      summarise(core_enrichment = paste(core_enrichment, collapse="/")) %>% 
      arrange(pvalue, p.adjust) 
      # Append description
      res.msigdf.em2 <- left_join(res.msigdf.em2, m_t2d, by = c("ID" = "gs_name"))
      res.msigdf.em2$Description <- res.msigdf.em2$gs_description

  res.msigdf.all <- c(res.msigdf.all, list(res.msigdf.em), list(res.msigdf.em2))
  names(res.msigdf.all)[length(res.msigdf.all) - 1] <- paste0("Enrich.", gs_cat)
  names(res.msigdf.all)[length(res.msigdf.all) - 0] <- paste0("GSEA.", gs_cat)
}
```

# Save the data

```{r}
x <- c(list(res.kegg), list(res.kegg.gsea), res.msigdf.all)
names(x)[1:2] <- c("Enrich.KEGG", "GSEA.KEGG")
write_xlsx(x, path = fileNameOut1)
```


```{r eval=FALSE}
# Custom enrichment

# More precise p-values cannot be extracted.

DREAM_signature <- read_xlsx(fileNameIn1, sheet = 3, col_names = FALSE)
DREAM_signature <- as.character(DREAM_signature$...1)
# Convert symbols to entrezids
eid <- bitr(DREAM_signature, fromType="SYMBOL", toType="ENTREZID", OrgDb=OrgDb)
TERM2GENE <- data.frame(ID = "DREAM", Gene = eid$ENTREZID)

em2 <- GSEA(geneList = geneList, TERM2GENE = TERM2GENE, pvalueCutoff = 1, maxGSSize = 8000)
kable(em2[, 1:10])
print(paste0("p-value:", em2@result$pvalue))
print(paste0("p-adjusted:", em2@result$p.adjust))
```

```{r fig.width = 6, fig.height = 3, eval=FALSE}
library(HGNChelper)
current_map <- getCurrentHumanMap()
library(msigdbr)
library(KEGGREST)
gs_cat <- "KEGG"
gs_selected <- "Ribosome"
query <- keggGet("hsa03010")
gs_selected <- "Oxidative phosphorylation"
query <- keggGet("hsa00190")
query_gene <- query[[1]]$GENE[grepl("^[0-9]", query[[1]]$GENE, perl = TRUE)]
# query_gene <- sapply(query_gene, function(x) strsplit(x, ";")[[1]][1])
# genes_checked <- checkGeneSymbols(query_gene, map = current_map)
# query_gene <- genes_checked$Suggested.Symbol
m_t2g <- data.frame(gs_name = gs_selected, entrez_gene = query_gene)

gs_cat <- "C2"
gs_selected_all <- c("SANA_RESPONSE_TO_IFNG_UP", "ONDER_CDH1_TARGETS_2_DN", "SANA_TNF_SIGNALING_UP", "JAEGER_METASTASIS_DN")

gs_cat <- "C5"
gs_selected_all <- c("GOMF_STRUCTURAL_CONSTITUENT_OF_RIBOSOME", "GOBP_DEFENSE_RESPONSE")

gs_cat <- "H"
gs_selected_all <- c("HALLMARK_INTERFERON_ALPHA_RESPONSE", "HALLMARK_INTERFERON_GAMMA_RESPONSE", "HALLMARK_WNT_BETA_CATENIN_SIGNALING", "HALLMARK_OXIDATIVE_PHOSPHORYLATION", "HALLMARK_DNA_REPAIR", "HALLMARK_NOTCH_SIGNALING", "HALLMARK_KRAS_SIGNALING_UP", "HALLMARK_MTORC1_SIGNALING")

for (gs_selected in gs_selected_all) {
  print(gs_selected)
  # gs_selected <- "HALLMARK_NOTCH_SIGNALING"
  m_t2g <- msigdbr(species = "Homo sapiens", category = gs_cat) %>% dplyr::select(gs_name, entrez_gene) %>% dplyr::filter(gs_name == !!gs_selected)
  em2 <- GSEA(geneList, TERM2GENE = m_t2g, pvalueCutoff = p_adj_cutoff)
  
  gseaplot(em2, geneSetID = 1, by = "runningScore") + ggtitle(gs_selected)
  ggsave(filename = file.path(data_dir, paste0("results/Figures_GSEA/Figure_GSEA_", gs_cat, "_", gs_selected, ".svg")), width = 4, height = 2.6, units = "in")
}
```
